+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Fix CommentedCodeCheck, splitting commented code on newlines.
 - Fix CommentedCodeCheck, not seeing seeing the commented line if line is top of file.
+- Fix UnusedVariableCheck, not handling try/when constructs properly.
 
 0.10.0 (2024-07-28)